backend/requirements.txt
fastapi
uvicorn
volcengine
python-multipart
python-dotenv
requests